created advertisement upload backend

+++ database/migrations/2023_10_25_165407_create_table_advertisements.php
@@ -22,7 +22,7 @@ return new class extends Migration
             $table->text('description');
             $table->string('subject');
             $table->boolean('status')->default(0); // You can set a default status value.
-            $table->unsignedBigInteger('tutor_id');
+            $table->unsignedInteger('tutor_id');
             $table->foreign('tutor_id')->references('id')->on('tutor_registers')->onUpdate('cascade')->onDelete('cascade');
 
             $table->timestamps();
